Converted landing page into home page
